backend/requirements.txt
Django==4.1.7
django-colorfield==0.8.0
django-filter==22.1
djangorestframework==3.14.0
djoser==2.1.0
drf-extra-fields==3.4.1
gunicorn==20.1.0
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
PyJWT==2.6.0
requests==2.28.2
